requirements.txt
pydub
openai-whisper
pysrt
numpy
librosa
transformers
accelerate
datasets
torch
torchinfo
soundfile

